\name{animation}
\alias{ex_animationMeanMedian}
\title{Animation of a content}

\description{Animation-example to show the usage of the functions of the package TGUICore. The examples can be called from the TGUI-interface.}
\usage{ex_animationMeanMedian()}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}
\keyword{methods}