\name{DATA_Collection}
\alias{ex_dataCollection}
\title{Mask for Data Collection}
\description{Example showing interactive student data input using a data-input-mask provided by package TGUICore. The mask can be called from the TGUI-interface.}
\usage{ex_dataCollection()}
\details{
  The TGUI provides a mask for data input. With it it's possible to include input from students directly into course examples, 
  course slides, graphics or tables. Different questions about current affairs or other course related things can be prepared in advance
  and asked during classroom teaching.
  The data mask in TGUITeaching provides an data collecting example with six different questions. 
  By default a fixed dummy-dataset is generated. The predefined dummy-data will be overwritten by student data input.  
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}

\keyword{Data Collection}
\keyword{Mask for Data Input}