\name{ex_BinomPlot}
\alias{ex_BinomPlot}
\title{Interactive Example}

\description{Interactive example to show the usage of the functions of the package TGUICore. The examples can be called with the GUI.}
\usage{ex_BinomPlot()}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}
\note{
Some visualisations are modified functions from the package TeachingDemos.
}
\keyword{methods}