\name{ex_praesident}
\alias{ex_praesident}
\alias{ex_animationMeanMedian}
\alias{ex_BinomPlot}
\alias{ex_boxplots}
\alias{ex_clt}
\alias{ex_dataCollection}
\alias{ex_eda}
\alias{ex_errors}
\alias{ex_feedback}
\alias{ex_Hists}
\alias{ex_QQPlots}
\alias{ex_timeseries}
\title{ex_praesident}
\description{Two examples fot the usage of the functionality of the package TGUICore}
\usage{ex_praesident(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example (if it exists) should be opened}
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlind Dinges}
\keyword{methods}