\name{exercise1}
\alias{exercise1}
\alias{feedback}
\title{exercise1}
\description{Two examples fot the usage of the functionality of the package TGUICore}
\usage{feedback(eval=FALSE)}
\usage{exercise1(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example should be opened}
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlind Dinges}
\keyword{methods}