\name{ex_feedback}
\alias{ex_feedback}
\title{Feature for Textual Feedback}

\description{Example to show the usage of the textual feedback feature of the package TGUICore. The feature can be called with the GUI.}
\usage{ex_feedback(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example (if it exists) should be opened}
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}

\keyword{methods}