\name{Simple_Examples}
\alias{ex_praesident}
\alias{ex_errors}
\title{Simple Generic Examples}

\description{Different simple generic examples to show the usage of the functions of the package TGUICore. The examples be called from the TGUI-interface.}
\usage{ex_praesident(eval=FALSE)}
\usage{ex_errors(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example (if it exists) should be opened}
}
\details{
  This two simple examples shows the use of single choice questions and multiple choice questions and how an image or a function can be included.  
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}

\keyword{Examples}