\name{TGUITeaching-package}
\alias{TGUITeaching-package}
\alias{TGUITeaching}
\docType{package}
\title{
Prototyp to enhance classroom teaching.
}
\description{
Teaching-GUI prototype, for creating TGUI content to enhance classroom teaching supported by TGUI. A Graphical User Interface to enhance and support modern teaching methods in an interactive way. 
This package uses TGUICore which provides all facilities needed to create interactive examples or ask for student feedback to enrich classroom teaching.  
It provides a student interface and a trainer interface and can be used in a classroom to train (statistical) contents or demonstrate (statistical) concepts.
Some generic example are included to show how exercises, interactive plots and animations can be generated or integrated.}
\details{
  \tabular{ll}{
    Package: \tab TGUITeaching\cr
    Type: \tab Package\cr
    Version \tab 9.8\cr
    Date: \tab 2010-03-23\cr
  }
  
  Examples in this package:

  \tabular{ll}{
    ex_errors, ex_praesident \tab Simple Generic Examples \cr
    ex_timeseries, ex_boxplots \tab Interactive Generic Examples \cr
    ex_ex_Hists, ex_QQPlots \tab Interactive Generic Examples \cr
    ex_BinomPlot, ex_clt, ex_eda \tab Interactive Generic Examples \cr
    ex_animation \tab Animation of a teaching content \cr
    ex_dataCollection \tab Mask for Data Collection \cr
    ex_feedback \tab Feature for Textual Feedback \cr
  }

}
\author{
  Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges
}
\keyword{package}
\seealso{ The TeachingDemos package }


