\name{ex_dataCollection}
\alias{ex_dataCollection}
\title{Mask for Data Collection}
\description{Example showing an example using interactive student data input using a data-input-mask provided by package TGUICore. The mask can be called from the TGUI-interface.}
\usage{ex_dataCollection()}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}

\keyword{methods}