\name{ex_praesident}
\alias{ex_praesident}
\alias{ex_animationMeanMedian}
\alias{ex_BinomPlot}
\alias{ex_boxplots}
\alias{ex_clt}
\alias{ex_dataCollection}
\alias{ex_eda}
\alias{ex_errors}
\alias{ex_feedback}
\alias{ex_Hists}
\alias{ex_QQPlots}
\alias{ex_timeseries}
\title{ex_praesident}
\description{Different examples for the usage of the functions of the package TGUICore. The examples can be called with the GUI.}
\usage{ex_praesident(eval=FALSE)}
\usage{ex_animationMeanMedian(eval=FALSE)}
\usage{ex_BinomPlot(eval=FALSE)}
\usage{ex_boxplots(eval=FALSE)}
\usage{ex_clt(eval=FALSE)}
\usage{ex_dataCollection(eval=FALSE)}
\usage{ex_eda(eval=FALSE)}
\usage{ex_errors(eval=FALSE)}
\usage{ex_feedback(eval=FALSE)}
\usage{ex_Hists(eval=FALSE)}
\usage{ex_QQPlots(eval=FALSE)}
\usage{ex_timeseries(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example (if it exists) should be opened}
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlind Dinges}
\keyword{methods}