\name{Feedback_Feature}
\alias{ex_feedback}
\title{Feature for Textual Feedback}

\description{Example to show the usage of the textual feedback feature of the package TGUICore. The feature be called from the TGUI-interface.}
\usage{ex_feedback(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example (if it exists) should be opened}
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}

\keyword{Feedback Feature (textual)}

