\name{Interactive_Examples}
\alias{ex_boxplots}
\alias{ex_clt}
\alias{ex_eda}
\alias{ex_Hists}
\alias{ex_QQPlots}
\alias{ex_timeseries}
\alias{ex_BinomPlot}
\title{Interactive Generic Examples}

\description{Different interactive examples to show the usage of the functions of the package TGUICore. The examples can be called from the TGUI-interface.}
\usage{ex_boxplots(eval=FALSE)}
\usage{ex_clt(eval=FALSE)}
\usage{ex_eda(eval=FALSE)}
\usage{ex_Hists(eval=FALSE)}
\usage{ex_QQPlots(eval=FALSE)}
\usage{ex_timeseries(eval=FALSE)}
\usage{ex_BinomPlot(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example (if it exists) should be opened}
}
\details{
  The different interactive examples shows the use of e.g. sliders and several interactive features.  
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}
\note{
Some visualisations are modified functions from the R package TeachingDemos.
}
\keyword{Examples}
\seealso{ The TeachingDemos package }