\name{ex_errors}
\alias{ex_praesident}
\alias{ex_errors}
\title{Simple Generic Examples}

\description{Different simple generic examples to show the usage of the functions of the package TGUICore. The examples can be called with the GUI.}
\usage{ex_praesident(eval=FALSE)}
\usage{ex_errors(eval=FALSE)}
\arguments{
	\item{eval}{TRUE/FALSE if the evaluation view of the example (if it exists) should be opened}
}
\author{Alexander Kowarik, Bernhard Meindl, Matthias Templ, Gerlinde Dinges}

\keyword{methods}